refactor: Rename component
